next version of protocol, try to deal with sockets more properly
The idea is that QDomDocument relies on reading exactly one document
from an io device or buffer. If the io device or buffer contains
more than one document, it gets lost.

On top of which, the "setContent" function is not reentrant, so it can't be
used directly with sockets.

AND, we can't really rely on a socket delivering one message at a time, it may
deliver more. Trying to throttle the sender is probably not reliable, so
consequently we make our own parser for messages.