seems ok now, but see below
it seems the heartbeat message has a problem (disabled for now),
AND that stops the file transfer in the middle.
Also after that, some messages get lost.

Have to test the recovery mechanism, if a messages is incomplete, do we
every try to re-parse it. And if so, at what point do we give up
and re-sync.